+++ src/main/java/com/gmail/val59000mc/scoreboard/ScoreboardManager.java
@@ -193,4 +193,9 @@ public class ScoreboardManager {
 	public void updatePlayerOnTab(UhcPlayer uhcPlayer) {
 		scoreboardHandler.updatePlayerOnTab(uhcPlayer);
 	}
+
+	public void updateTeamOnTab(UhcTeam uhcTeam) {
+		scoreboardHandler.updateTeamOnTab(uhcTeam);
+	}
+
 }